initial separation
