+++ airbyte-integrations/connectors/source-alloydb-strict-encrypt/src/main/java/io/airbyte/integrations/source/alloydb/AlloyDbStrictEncryptSource.java
@@ -8,7 +8,6 @@ import static io.airbyte.integrations.source.relationaldb.state.StateManager.LOG
 
 import io.airbyte.integrations.base.IntegrationRunner;
 import io.airbyte.integrations.base.Source;
-import io.airbyte.integrations.source.postgres.PostgresSourceStrictEncrypt;
 
 public class AlloyDbStrictEncryptSource {
