+++ airbyte-integrations/connectors/source-postgres-strict-encrypt/src/main/java/io/airbyte/integrations/source/postgres_strict_encrypt/PostgresSourceStrictEncrypt.java
@@ -2,25 +2,28 @@
  * Copyright (c) 2023 Airbyte, Inc., all rights reserved.
  */
 
-package io.airbyte.integrations.source.postgres;
+package io.airbyte.integrations.source.postgres_strict_encrypt;
+
+import static io.airbyte.protocol.models.v0.AirbyteConnectionStatus.Status;
 
 import com.fasterxml.jackson.databind.JsonNode;
+import com.fasterxml.jackson.databind.node.ArrayNode;
 import com.fasterxml.jackson.databind.node.ObjectNode;
 import io.airbyte.commons.json.Jsons;
 import io.airbyte.db.jdbc.JdbcUtils;
 import io.airbyte.integrations.base.IntegrationRunner;
 import io.airbyte.integrations.base.Source;
 import io.airbyte.integrations.base.spec_modification.SpecModifyingSource;
+import io.airbyte.integrations.source.postgres.PostgresSource;
 import io.airbyte.protocol.models.v0.AirbyteConnectionStatus;
-import io.airbyte.protocol.models.v0.AirbyteConnectionStatus.Status;
 import io.airbyte.protocol.models.v0.ConnectorSpecification;
 import java.util.Set;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 /**
- * This file will soon be removed. Any change to this file should also be duplicated to
- * PostgresSourceStrictEncrypt.java in the source-postgres module.
+ * This class is copied from source-postgres-strict-encrypt. The original file can be deleted
+ * completely once the migration of multi-variant connector is done.
  */
 public class PostgresSourceStrictEncrypt extends SpecModifyingSource implements Source {
 
@@ -34,17 +37,20 @@ public class PostgresSourceStrictEncrypt extends SpecModifyingSource implements 
   public static final String SSL_MODE_DISABLE = "disable";
   public static final String SSL_MODE_REQUIRE = "require";
 
-  PostgresSourceStrictEncrypt() {
+  public PostgresSourceStrictEncrypt() {
     super(PostgresSource.sshWrappedSource());
   }
 
   @Override
   public ConnectorSpecification modifySpec(final ConnectorSpecification originalSpec) {
     final ConnectorSpecification spec = Jsons.clone(originalSpec);
-    final ObjectNode properties = (ObjectNode) spec.getConnectionSpecification().get("properties");
-    properties.remove(JdbcUtils.SSL_KEY);
-    ((ObjectNode) properties.get(SSL_MODE)).put("default", SSL_MODE_REQUIRE);
-
+    ((ObjectNode) spec.getConnectionSpecification().get("properties")).remove(JdbcUtils.SSL_KEY);
+    final ArrayNode modifiedSslModes = spec.getConnectionSpecification().get("properties").get("ssl_mode").get("oneOf").deepCopy();
+    // Assume that the first item is the "disable" option; remove it
+    modifiedSslModes.remove(0);
+    ((ObjectNode) spec.getConnectionSpecification().get("properties").get(SSL_MODE)).remove("oneOf");
+    ((ObjectNode) spec.getConnectionSpecification().get("properties").get(SSL_MODE)).set("oneOf", modifiedSslModes);
+    ((ObjectNode) spec.getConnectionSpecification().get("properties").get(SSL_MODE)).put("default", SSL_MODE_REQUIRE);
     return spec;
   }
 
@@ -75,5 +81,4 @@ public class PostgresSourceStrictEncrypt extends SpecModifyingSource implements 
     new IntegrationRunner(source).run(args);
     LOGGER.info("completed source: {}", PostgresSourceStrictEncrypt.class);
   }
-
 }
