+++ airbyte-integrations/connectors/source-postgres-strict-encrypt/src/test-integration/java/io/airbyte/integrations/source/postgres_strict_encrypt/PostgresSourceStrictEncryptAcceptanceTest.java
@@ -1,0 +1,5 @@
+package io.airbyte.integrations.source.postgres_strict_encrypt;
+
+public class PostgresSourceStrictEncryptAcceptanceTest {
+
+}
