+++ airbyte-integrations/connectors/source-postgres-strict-encrypt/src/test-integration/resources/dummy_config.json
@@ -1,0 +1,6 @@
+{
+  "host": "default",
+  "port": 5555,
+  "database": "default",
+  "username": "default"
+}
